ai-engine/requirements.txt
# Context Cloud — Python AI Engine
# Core ML/NLP dependencies
fastapi==0.115.6
uvicorn[standard]==0.32.1
grpcio==1.68.1
grpcio-tools==1.68.1
protobuf==5.29.2

# NLP & NER
spacy==3.8.3

# Embeddings
sentence-transformers==3.3.1
torch==2.5.1

# Data validation
pydantic==2.10.3

# Utilities
python-dotenv==1.0.1
structlog==24.4.0
